+++ samples/walletg.py
@@ -6,6 +6,7 @@
 """Main driver primarily for demonstrating using Sui GraphQL."""
 
 import argparse
+import asyncio
 import os
 import pathlib
 import sys
@@ -17,17 +18,15 @@ sys.path.insert(0, str(PROJECT_DIR))
 sys.path.insert(0, str(PARENT))
 sys.path.insert(0, str(os.path.join(PARENT, "pysui")))
 
-from pysui import PysuiConfiguration, SyncGqlClient
-
-# from pysui.sui.sui_pgql.pgql_clients import SuiGQLClient
+from pysui import PysuiConfiguration, client_factory
 from pysui.sui.sui_constants import PYSUI_CLIENT_CONFIG_ENV
 
 from samples.cmd_argsg import build_parser, pre_config_pull
 from samples.cmdsg import SUI_CMD_DISPATCH
 
 
-def main():
-    """Entry point for demonstration."""
+async def _run():
+    """Async entry point."""
     cfg, arg_line = pre_config_pull(sys.argv[1:].copy())
     parsed = build_parser(arg_line, cfg)
     cmd_call = SUI_CMD_DISPATCH.get(parsed.subcommand, None)
@@ -38,9 +37,14 @@ def main():
         var_args = vars(parsed)
         var_args.pop("subcommand")
         parsed = argparse.Namespace(**var_args)
-        cmd_call(SyncGqlClient(pysui_config=cfg), parsed)
+        await cmd_call(client_factory(cfg), parsed)
     else:
         print(f"Unable to resolve function for {parsed.subcommand}")
+
+
+def main():
+    """Entry point for demonstration."""
+    asyncio.run(_run())
 
 
 if __name__ == "__main__":